u-boot: Fix UBOOT_INITIAL_ENV
fw_printenv and Swupdate expect the initial environment to be placed
at /etc/u-boot-initial-env, but currently it is being placed at
/etc/u-boot-mainline-initial-env instead.

This happens due to the usage of the {PN} notation, which pulls the
'u-boot-mainline' string.

Remove the {PN} notation and pass the full file name to fix the problem.

Signed-off-by: Fabio Estevam <festevam@denx.de>